add ST[1,6] to intermidiate steps and make stepId uint16

+++ inc/Helpers/TypeDefinitions.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 
-using StepId = uint8_t;
+using StepId = uint16_t;
 using ParameterReportStructureId = uint16_t;
 using EventDefinitionId = uint16_t;
 using EventActionId = uint16_t;
@@ -20,7 +20,7 @@ using MemoryDataLength = uint16_t;
 /**
  * Parameter Monitoring Identifier as used in the On-board Monitoring Service.
  */
-using LargeMessageTransactionId = uint16_t;
+using LargeMessageTransactionId = uint8_t;
 using PartSequenceNum = uint16_t;
 using PacketStoreSize = uint16_t;
 using PacketStoreType = uint8_t;
